requirements.txt
fastapi==0.115.12
numpy==2.2.6
opencv-python-headless==4.10.0.84
pillow==11.1.0
pytest==8.3.5
python-multipart==0.0.20
uvicorn[standard]==0.34.0
